requirements.txt
python-telegram-bot==20.7
sqlalchemy
databases
psycopg2-binary   # for Postgres on Render
